♻️ refactor: refactor the globalStore implement and usage
