+++ src/store/global/index.ts
@@ -1,3 +1,2 @@
 export * from './hooks';
-export * from './selectors';
 export * from './store';
