+++ src/store/global/slices/settings/action.ts
@@ -5,12 +5,12 @@ import { DeepPartial } from 'utility-types';
 import type { StateCreator } from 'zustand/vanilla';
 
 import { DEFAULT_AGENT, DEFAULT_SETTINGS } from '@/const/settings';
+import type { GlobalStore } from '@/store/global';
 import { SettingsTabs } from '@/store/global/initialState';
+import { LobeAgentSettings } from '@/types/session';
 import type { GlobalSettings, OpenAIConfig } from '@/types/settings';
 import { merge } from '@/utils/merge';
 import { setNamespace } from '@/utils/storeDebug';
-
-import type { GlobalStore } from '../store';
 
 const n = setNamespace('settings');
 
@@ -24,6 +24,7 @@ export interface SettingsAction {
    */
   resetSettings: () => void;
   setOpenAIConfig: (config: Partial<OpenAIConfig>) => void;
+
   /**
    * 设置部分配置设置
    * @param settings - 部分配置设置
@@ -35,6 +36,7 @@ export interface SettingsAction {
    * @param themeMode - 主题模式
    */
   switchThemeMode: (themeMode: ThemeMode) => void;
+  updateDefaultAgent: (agent: DeepPartial<LobeAgentSettings>) => void;
 }
 
 export const createSettingsSlice: StateCreator<
@@ -78,8 +80,14 @@ export const createSettingsSlice: StateCreator<
   switchSettingTabs: (tab) => {
     set({ settingsTab: tab });
   },
-
   switchThemeMode: (themeMode) => {
     get().setSettings({ themeMode });
+  },
+  updateDefaultAgent: (agent) => {
+    const settings = produce(get().settings, (draft: GlobalSettings) => {
+      draft.defaultAgent = merge(draft.defaultAgent, agent);
+    });
+
+    set({ settings }, false, n('updateDefaultAgent', agent));
   },
 });
