+++ src/store/session/slices/agent/selectors.ts
@@ -4,7 +4,8 @@ import { t } from 'i18next';
 import { DEFAULT_OPENAI_MODEL_LIST, VISION_MODEL_WHITE_LIST } from '@/const/llm';
 import { DEFAULT_AVATAR, DEFAULT_BACKGROUND_COLOR, DEFAULT_INBOX_AVATAR } from '@/const/meta';
 import { DEFAULT_AGENT_CONFIG, DEFAUTT_AGENT_TTS_CONFIG } from '@/const/settings';
-import { settingsSelectors, useGlobalStore } from '@/store/global';
+import { useGlobalStore } from '@/store/global';
+import { settingsSelectors } from '@/store/global/selectors';
 import { SessionStore } from '@/store/session';
 import { LobeAgentTTSConfig } from '@/types/agent';
 import { LanguageModel } from '@/types/llm';
