+++ src/store/session/slices/session/action.ts
@@ -5,7 +5,8 @@ import { StateCreator } from 'zustand/vanilla';
 import { INBOX_SESSION_ID } from '@/const/session';
 import { SESSION_CHAT_URL } from '@/const/url';
 import { sessionService } from '@/services/session';
-import { settingsSelectors, useGlobalStore } from '@/store/global';
+import { useGlobalStore } from '@/store/global';
+import { settingsSelectors } from '@/store/global/selectors';
 import { SessionStore } from '@/store/session';
 import {
   LobeAgentSession,
